CssCompositeValueNode does not perform deepCopy correctly. This becomes an issue when we allowed the compiler to understand mixin parameters in a CssCompositeValueNode.
The problem manifests when a CssCompositeValueNode's children are nested.
The constant references in those children will be modified in the original
mixin definitions instead of the cloned subtree.

The unit test shows a concrete example of this issue.
	Change on 2012/05/01 by chrishenry@chrishenry:closure_mini_3:11:citc
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=29729263