Remove 'addScalar'/'subScalar' methods, fixing the bug introduced in August last year by [] which incorrectly made 'add' and 'sub' require units when evaluated as a string.
In doing so remove some overrides. Note, the bug fixed by that CL still has
its unit tests in place (and is verified to still work); except for the new
test which incorrectly asserted unitless addition/subtraction was not permissible.

As the number of clients of the *Scalar methods was small, remove them in
this same CL.
	Change on 2013/07/25 by nmvc <nmvc@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=49906955