Part 1 of support for CSS selector auto-prefixing. This modifies the @component rule to make the component name optional, in which case the prefix for class names and def constants is derived from the @provide package name. This also includes changes to the Quark appbar showing how the new feature can be used.
Note that there's one oddness here, which is that I had to use
SourceLocation (which is normally used only for debugging) to be able
to match @component rules with the correct @provide rule. The reason
for this is (a) at this point in the processing all of the GSS files
have been concatenated together, so that there are many @component /
@provide directives mixed together in the same scope, and (b) the tree
visitor interface gives no means to detect the boundaries between
source files. I'm open to suggestions on how to do this better.

Other than that the change is relatively straightforward.

This is a second try of CL 67707321 that was rolled back in CL
67712609. This changes the logic so that multiple @provides are once
again allowed unless there is an implicitly-named component.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=67768691