Part 2 of CSS package-relative selector work. This adds support for the '%' prefix to opt-in to prefixing. It also updates the logic to only prefix the first class selector in a selector expression.
The new behavior is disabled via a flag. This means that the the old
behavior, which is to prefix every class selector, will be in
effect. There are unit tests for both the old and new behavior.

Even when disabled, the new % syntax is accepted, however it has no
effect. This will allow us to add the % syntax to existing GSS files
in preparation for enabling the new behavior.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=68913910