This CL introduces SimpleSplittingSubstitutionMap to allow easier debugging. This is a simple combination of the SimpleSubstitutionMap and SplittingSubstitutionMap.
The reason why this is necessary is because
CLOSURE_COMPILED_SPLIT_HYPHENS is really the only
--output_renaming_map_format that *fully* works, based on my
testing. But there doesn't seem to be a way to use the SPLIT_HYPHENS
mode so that the class renames are still human-readable (with the
trailing underscore) for debugging purposes. These classes solve
this problem:

  --output_renaming_map_format=CLOSURE_COMPILED_SPLIT_HYPHENS
  --css_substitution_map_provider=com.google.common.css.SimpleSplittingSubstitutionMapProvider

  goog.setCssNameMapping({
    "kd": "kd_",
    "button": "button_",
  });

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=78028032