Fix bug in CSS parser which produces a NullPointerException when error recovery is enabled.
The problem is that when the normal constructor is used for
ParseException, e.currentToken is set to null. The error handling
logic advances the parser forward, but because e.currentToken is null
it rethrows the Exception all the way back up to parse(). parse()
catches the ParseException and attempts to convert it to
GssParserException, but because token.next is null it throws a NPE
while trying to get the location.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=85444801