Add css calc() function to auto expanding browser prefixes.
This involves adapting the auto expansion code to handle rules that have
no specific property name associated with them but apply to the value
anywhere they are seen. This is because calc() can be used anywhere a
<length>, <frequency>, <angle>, <time>, <number>, or <integer> is
required, and is not property specific.

This change also allows for more than one value to be expanded in the
same declaration, since previously only the first value could be
matched. This is needed when the expansion value can appear twice in the
values, such as: margin: calc(50% - 10px) calc(100% + 20px).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=99542674