Fix lack of SourceCodeLocation in @components.
There is an issue where DisallowDuplicateDeclarations pass fails
pre-condition check due to the ruleset having a null
SourceCodeLocation. This can happen in rulesets in @component.  This
change improves the situation by using the SourceCodeLocation from the
first selector in the ruleset, if the location is null in the ruleset.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=101045499