Improve SourceCodeLocation accuracy.
CssDeclarationNode did not have the SourceCodeLocation as a constructor
parameter, nor did it preserve it on a deep copy.

CssNodesListNode now tries to build its locations from its constituent nodes
unless a SourceCodeLocation is set explicitly.

Several bugs in the actual parser are also fixed as are numerous problems in
FixupFontDeclarations.

There are new tests and a SourceCodeLocationSubject for Truth which should
make new tests easier to write.