Add --input-renaming-map and --input-renaming-map-format options
With these options, it is possible to provide a file
that contain a renaming map in command line.

This feature is useful to build several css files
with the same map (like themes, specific devices, ...).

The generated map contains the content of the input map and
others renaming fields that did not in input map.