Add --input-renaming-map and --input-renaming-map-format options
With these options, it is possible to provide a file
that contains a renaming map on the command line.

This feature is useful to build several css files
with the same map (like themes, specific devices, ...).

The generated map contains the content of the input map and
the renamed classes that where not in the input map.

Closes GitHub issue #8.

Contributed by Sébastien Mennetrier <smennetrier@voxtok.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=148128344