Fix VisitCompositeValueState.replaceCurrentBlockChildWith
Before, the method would not set the parent for the children. Now it
does. As far as I can tell, this was never a major problem. But it was
making debugging more difficult and it's pretty inexcusable that we
don't adhere to our contract. (I can say that as the author of the
broken code. :-)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=149427832