Adds basic support for CSS custom properties [https://drafts.csswg.org/css-variables]
Custom property values are parsed as CSS expressions, which differs
from the spec. The spec is very permissive -- its only requirements
being matched braces/parens/brackets, and terminated strings. We have
the option of expanding to this definition in the future, but for now
restrictive values are a safer and simpler option.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158885623