Annotate tree visitor `enterXXX` methods with @CheckReturnValue
Also, refactor DefaultVisitController to make it easier to check the
enterXXX() return values correctly.

GSS' AtRuleHandler and CssTreeVisitor interfaces have enter() and
leave() methods for each kind of node in the AST. While the leave()
methods are void, the enter() methods return a boolean. The contract
is that if an implementation returns false, the node's children should
not be visited.

Ignoring the return value from an enter() method is always wrong,
since it can cause parts of the tree to be visited that the callee did
not want visited, or vice versa. (we found an example of code that
unconditionally visited a subtree, ignoring the callee's intent.) This
CL annotates all the enter() methods with @CheckReturnValue, making
such bugs compilation errors.

The inner classes of DefaultVisitController are the main
implementations of both interfaces. Currently, the inner classes use
ad-hoc logic to propagate the enter() return values correctly. (For
example, several use two booleans, `visitedChildren` and
`shouldVisitChildren`.) This CL introduces an enum,
VisitState.Progress, to make the flow of control less error-prone.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158975760