Replace truth.FailureStrategy with truth.FailureMetadata in custom Subjects.
Also changed truth.SubjectFactory to truth.Subject.Factory (plain
renaming) and use method reference instead of anonymous class to
create the factory when applicable.

FailureMetadata, an opaque object to its users, is introduced to
replace FailureStrategy in in custom Subject in order to resolve some
existing flaws of FailureStrategy as well as enable new features to be
added to Truth.

New API is available in Truth-0.36, if there is a build/pom.xml, it's
also updated to use this version.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175823335